requirements.txt
fastapi==0.115.0
sentry-sdk[fastapi]==2.19.2
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.3.4
sqlalchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
itsdangerous==2.2.0
httpx==0.27.0
email-validator==2.1.1
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
gunicorn==21.2.0
python-dotenv==1.0.0
aiosmtplib==3.0.0

# Rate limiting
slowapi==0.1.9

# Reporting dependencies
openpyxl==3.1.5
reportlab==4.4.4
matplotlib==3.10.6
seaborn==0.13.2
pandas==2.2.3

# Machine Learning dependencies
scikit-learn==1.5.0
joblib==1.4.2
numpy==1.26.4
prophet==1.1.5
cmdstanpy==1.2.0

# Background task scheduling
apscheduler==3.10.4

# OCR – Receipt scanning (Phase 32 Part B)
# Requires Tesseract binary: apt-get install tesseract-ocr (Railway/Linux) or https://github.com/UB-Mannheim/tesseract/wiki (Windows)
pytesseract==0.3.13
Pillow==10.4.0

# Caching (optional - for performance optimization)
redis==5.0.1

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.27.0
psutil==5.9.6